Add notes for most of section 9
I got half way through the intermediate scripting section before I gave
up. I have enough of the basics of JavaScript. Now, I need to focus on
the DOM. Remember that the point of this is to bodge a client-side JS
content manager that grabs, translates, and exposes the markdown I have
on the server side.

Signed-off-by: dafydd2277 <dafydd2277@gmail.com>